move progressBar to objects/progress.py
+ move progressBar class definition from utils/ptime.py to objects/progress.py for simplicity. Old usage of ptime.progressBar stays the same.

+ smallbaselineApp.py -H now printout syntax highlighed text if rich module is installed.
