+++ mintpy/objects/__init__.py
@@ -1,16 +1,3 @@
 from .giant import *
 from .stack import *
 from .ramp import *
-
-
-## Sub-module dependency graph:
-## Check mintpy/__init__.py for more comprehensiave graph
-# Level 0 modules:
-# giant
-# stack
-# ramp
-# sensor
-# colors
-# Level 1 modules:
-# conncomp (objects/ramp)
-#
